refactor: domain 모듈의 패키지 관심사에 맞게 분리
