+++ data/src/main/java/camp/nextstep/edu/github/data/di/RepositoryModule.kt
@@ -3,7 +3,7 @@ package camp.nextstep.edu.github.data.di
 import camp.nextstep.edu.github.data.repository.DefaultGithubRepository
 import camp.nextstep.edu.github.data.service.GithubService
 import camp.nextstep.edu.github.data.entity.RepositoryMapper
-import camp.nextstep.edu.github.domain.GithubRepository
+import camp.nextstep.edu.github.domain.repository.GithubRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
