+++ data/src/main/java/camp/nextstep/edu/github/data/repository/DefaultGithubRepository.kt
@@ -2,8 +2,8 @@ package camp.nextstep.edu.github.data.repository
 
 import camp.nextstep.edu.github.data.service.GithubService
 import camp.nextstep.edu.github.data.entity.RepositoryMapper
-import camp.nextstep.edu.github.domain.GithubRepository
-import camp.nextstep.edu.github.domain.Repository
+import camp.nextstep.edu.github.domain.repository.GithubRepository
+import camp.nextstep.edu.github.domain.repository.Repository
 
 internal class DefaultGithubRepository(
     private val githubService: GithubService,
