+++ data/src/test/java/camp/nextstep/edu/github/data/repository/DefaultGithubRepositoryTest.kt
@@ -3,7 +3,7 @@ package camp.nextstep.edu.github.data.repository
 import camp.nextstep.edu.github.data.service.GithubService
 import camp.nextstep.edu.github.data.entity.RepositoryEntity
 import camp.nextstep.edu.github.data.entity.RepositoryMapper
-import camp.nextstep.edu.github.domain.Repository
+import camp.nextstep.edu.github.domain.repository.Repository
 import com.google.common.truth.Truth.assertThat
 import io.mockk.coEvery
 import io.mockk.coVerify
